Fix build issue caused by debugger class missing deps for release
