+++ app/src/main/kotlin/com/sniped/di/AppModule.kt
@@ -9,7 +9,6 @@ import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.sniped.BuildConfig
 import com.sniped.Debugger
-import com.sniped.DevDebugger
 import com.sniped.R
 import com.sniped.api.ApiService
 import com.sniped.api.ApiServiceBuilder
@@ -36,13 +35,11 @@ class AppModule(private val context: Context) {
     @Provides
     @Singleton
     fun provideApiService(
-            gson: Gson,
-            debugger: Debugger): ApiService {
+            gson: Gson): ApiService {
         val apiUri = context.getString(R.string.api)
         val apiServiceBuilder = ApiServiceBuilder(
                 apiUri,
-                gson,
-                debugger)
+                gson)
 
         return apiServiceBuilder
                 .build()
@@ -56,8 +53,4 @@ class AppModule(private val context: Context) {
     @Provides
     @Singleton
     fun provideSharedPreferences(): SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-
-    @Provides
-    @Singleton
-    fun provideDebugger() = if (BuildConfig.DEBUG) DevDebugger() else Debugger.NULL
 }
