+++ app/src/release/java/com/sniped/Debugger.kt
@@ -1,0 +1,11 @@
+package com.sniped
+
+import android.app.Application
+import okhttp3.OkHttpClient
+
+object Debugger {
+
+    fun attachToApp(app: Application) = Unit
+
+    fun attachToNetwork(okHttpClientBuilder: OkHttpClient.Builder) = Unit
+}
